move queue handler to services
